requirements.txt
# accelerate==1.2.0
# aiofiles==23.2.1
# aiohappyeyeballs==2.4.4
# aiohttp==3.11.10
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.7.0
# asttokens==3.0.0
# attrs==24.2.0
# backcall==0.2.0
# beautifulsoup4==4.12.3
# bitsandbytes==0.45.0
# bleach==6.2.0
# certifi==2024.8.30
# charset-normalizer==3.4.0
# click==8.1.7
# colorama==0.4.6
colpali_engine==0.3.4
# comm==0.2.2
# datasets==3.1.0
# debugpy==1.8.9
# decorator==5.1.1
# defusedxml==0.7.1
# dill==0.3.8
# diskcache==5.6.3
# docopt==0.6.2
docx2pdf==0.1.8
# easyocr==1.7.2
# executing==2.1.0
fastapi==0.115.6
# fastjsonschema==2.21.1
# ffmpy==0.4.0
# filelock==3.13.1
# frozenlist==1.5.0
# fsspec==2024.2.0
# GPUtil==1.4.0
gradio==5.8.0
gradio_client==1.5.1
# grpcio==1.68.1
# grpcio-tools==1.68.1
# h11==0.14.0
# h2==4.1.0
# hpack==4.0.0
# httpcore==1.0.7
# httpx==0.28.1
# huggingface-hub==0.26.3
# hyperframe==6.0.1
# idna==3.10
# imageio==2.36.1
# IProgress==0.4
# ipykernel==6.29.5
# ipython==8.12.3
# ipywidgets==8.1.5
# jedi==0.19.2
# Jinja2==3.1.4
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# jupyterlab_pygments==0.3.0
# jupyterlab_widgets==3.0.13
# lazy_loader==0.4
# llama_cpp_python==0.3.2
# lxml==5.3.0
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# mistune==3.0.2
# mpmath==1.3.0
# multidict==6.1.0
# multiprocess==0.70.16
# nbclient==0.10.1
# nbconvert==7.16.4
# nbformat==5.10.4
# nest-asyncio==1.6.0
# networkx==3.2.1
# ninja==1.11.1.2
# numpy==2.1.3
# opencv-python-headless==4.10.0.84
# orjson==3.10.12
# packaging==24.2
# pandas==2.2.3
# pandocfilters==1.5.1
# parso==0.8.4
# peft==0.11.1
# pickleshare==0.7.5
pillow==10.2.0
# platformdirs==4.3.6
# portalocker==2.10.1
# prompt_toolkit==3.0.48
# propcache==0.2.1
# protobuf==5.29.1
# psutil==6.1.0
# pure_eval==0.2.3
# pyarrow==18.1.0
# pyclipper==1.3.0.post6
pydantic==2.10.3
pydantic_core==2.27.1
# pydub==0.25.1
# Pygments==2.18.0
PyMuPDF==1.25.0
# PyPDF2==3.0.1
# pytesseract==0.3.13
# python-bidi==0.6.3
# python-dateutil==2.9.0.post0
# python-docx==1.1.2
# python-multipart==0.0.19
# pytz==2024.2
# pywin32==308
# PyYAML==6.0.2
# pyzmq==26.2.0
qdrant-client==1.12.1
# referencing==0.35.1
# regex==2024.11.6
# requests==2.32.3
# rich==13.9.4
# rpds-py==0.22.3
# ruff==0.8.2
# safehttpx==0.1.6
# safetensors==0.4.5
# scikit-image==0.24.0
# scipy==1.14.1
# semantic-version==2.10.0
# sentencepiece==0.2.0
# setuptools==75.6.0
# shapely==2.0.6
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# soupsieve==2.6
# stack-data==0.6.3
# stamina==24.3.0
# starlette==0.41.3
# sympy==1.13.1
# tenacity==9.0.0
# tifffile==2024.9.20
# timm==1.0.12
# tinycss2==1.4.0
# tokenizers==0.20.3
# tomlkit==0.13.2

bitsandbytes
# torchvision==0.20.1+cu124
# tornado==6.4.2
tqdm==4.67.1
# traitlets==5.14.3
# transformers==4.46.3
# typer==0.15.1
# typing_extensions==4.12.2
# tzdata==2024.2
# urllib3==2.2.3
# uv==0.5.6
uvicorn==0.32.1
# wcwidth==0.2.13
# webencodings==0.5.1
# websockets==14.1
# wheel==0.45.1
# widgetsnbextension==4.0.13
# xxhash==3.5.0
# yarg==0.1.9
# yarl==1.18.3
